// src/app/interfaces/common.interface.ts
export interface LoadingState {
    loading: boolean;
    loaded: boolean;
}

export interface QueryConfig {
    path: string; //  path to collection
    field: string; // field to orderBy
    limit: number; // limit per query
    reverse: boolean; // reverse order?
    prepend: boolean; // prepend to source?
  }